Crawl and store musical data directly on SQLite database
This program tracks.py reads file Library.xml which is attached here too. Creates 4 tables Album, Artist, Genre and Track. Opens the files which stores dictionary inside dictionary and reads the file and store the data directly in the created tables with respect ot their name.
The database is stored as trackdb.sqlite